stateful_complaint_chatbot/requirements.txt
streamlit==1.31.0
pandas==2.2.0
openpyxl==3.1.2
google-generativeai==0.3.2
python-dotenv==1.0.0
